% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ESV.R
\name{calculate_ESV}
\alias{calculate_ESV}
\title{Calculate Effective Spatial Variability}
\usage{
calculate_ESV(sigma.sq_vec, tau.sq, spatial_coords, phi_seq)
}
\arguments{
\item{sigma.sq_vec}{Vector of spatial variance components.}

\item{tau.sq}{Nugget variance component.}

\item{spatial_coords}{Spatial coordinates matrix.}

\item{phi_seq}{Sequence of length-scale parameters.}
}
\value{
Numeric value representing the ESV score (between 0 and 1).
}
\description{
Calculates the Effective Spatial Variability (ESV) score based on variance
components and spatial weights.
}
\keyword{internal}
