% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test.R
\name{score_test}
\alias{score_test}
\title{Score Test for Spatial Gene Expression}
\usage{
score_test(y, spatial_coords, phi_seq)
}
\arguments{
\item{y}{Gene expression vector.}

\item{spatial_coords}{Spatial coordinates matrix.}

\item{phi_seq}{Sequence of length-scale parameters.}
}
\value{
List containing p-values for each length scale.
}
\description{
Performs score tests to assess spatial gene expression patterns across
multiple length scales.
}
\keyword{internal}
