% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Visium_human_DLPFC}
\alias{Visium_human_DLPFC}
\title{Visium human DLPFC data}
\format{
\subsection{\code{Visium_human_DLPFC}}{

A list with 3 elements
\describe{
\item{counts}{Gene expression count matrix. Rows and columns indicate genes and spots, respectively.}
\item{spatial_coords}{Spatial coordinates matrix. Rows indicate spots.}
\item{cell_type_proportions}{Cell type proportion matrix estimated by RCTD. Rows and columns indicate spots and cell types, respectively.}
}
}
}
\source{
The Visium_human_DLPFC dataset contains
}
\usage{
Visium_human_DLPFC
}
\description{
An example dataset
}
\keyword{datasets}
