% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Visium_human_DLPFC}
\alias{Visium_human_DLPFC}
\title{Visium human DLPFC data}
\format{
\subsection{\code{Visium_human_DLPFC}}{

A list with 3 elements
\describe{
\item{counts}{\strong{\strong{}
\item{spatial_coords}{}}}
\item{cell_type_proportions}{****}
}
}
}
\source{
The Visium_human_DLPFC dataset contains
}
\usage{
Visium_human_DLPFC
}
\description{
An example dataset
}
\keyword{datasets}
